+++ sparrow-data/parse/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "sparrow-parse"
-version = "0.1.8"
+version = "0.1.9"
 description = "Sparrow Parse is a Python package for parsing and extracting information from documents."
 authors = ["Andrej Baranovskij <andrejus.baranovskis@gmail.com>"]
 license = "GPL-3.0"
@@ -20,8 +20,11 @@ include = [
 
 
 [tool.poetry.dependencies]
-python = "^3.10"
-requests = "^2.31.0"
+python = ">=3.9,<3.12"
+torch = {version = "2.2.2", source = "pypi"}
+unstructured = {version = "0.13.6", extras = ["all-docs"]}
+unstructured-inference = "0.7.29"
+rich = "^13.7.1"
 
 
 [tool.poetry.scripts]
