+++ sparrow-data/parse/sparrow_parse/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.1.8'
+__version__ = '0.1.9'
