--- sparrow-data/parse/sparrow_parse/pdf/pdf_processor.py
@@ -1,7 +1,0 @@
-class PDFProcessor(object):
-    def __init__(self):
-        pass
-
-    def process_file(self, content):
-        print("Processing file...")
-        return "OK"
