+++ sparrow-ml/llm/rag/agents/instructor/instructor.py
@@ -1,16 +1,10 @@
 from rag.agents.interface import Pipeline
 from openai import OpenAI
 import instructor
-from unstructured.partition.pdf import partition_pdf
-from unstructured.partition.image import partition_image
-from unstructured.staging.base import elements_to_json
+from sparrow_parse.extractor.file_processor import FileProcessor
 from pydantic import create_model
-# from sparrow_parse.pdf.pdf_processor import PDFProcessor
 from typing import List
 from rich.progress import Progress, SpinnerColumn, TextColumn
-import tempfile
-import os
-import json
 import timeit
 from rich import print
 from typing import Any
@@ -50,39 +44,11 @@ class InstructorPipeline(Pipeline):
         strategy = cfg.STRATEGY_INSTRUCTOR
         model_name = cfg.MODEL_INSTRUCTOR
 
-        # processor = PDFProcessor()
-        # result = processor.process_file(file_path, strategy, model_name)
+        processor = FileProcessor()
+        content = processor.extract_data(file_path, strategy, model_name, options, local, debug)
 
-        # check if string options contains word table
-        extract_tables = False
-        if options is not None and "tables" in options:
-            extract_tables = True
-
-        # Extracts the elements from the PDF
-        elements = self.invoke_pipeline_step(
-            lambda: self.process_file(file_path, strategy, model_name),
-            "Extracting elements from the document...",
-            local
-        )
-
-        if debug:
-            new_extension = 'json'  # You can change this to any extension you want
-            new_file_path = self.change_file_extension(file_path, new_extension)
-
-            content = self.invoke_pipeline_step(
-                lambda: self.load_text_data(elements, new_file_path, extract_tables),
-                "Loading text data...",
-                local
-            )
-        else:
-            with tempfile.TemporaryDirectory() as temp_dir:
-                temp_file_path = os.path.join(temp_dir, "file_data.json")
-
-                content = self.invoke_pipeline_step(
-                    lambda: self.load_text_data(elements, temp_file_path, extract_tables),
-                    "Loading text data...",
-                    local
-                )
+        # with open("data/invoice_1.txt", 'r') as file:
+        #     content = file.read()
 
         if debug:
             print(f"\nContent: {content}\n")
@@ -131,75 +97,6 @@ class InstructorPipeline(Pipeline):
         answer = resp.model_dump_json(indent=4)
 
         return answer
-
-    def process_file(self, file_path, strategy, model_name):
-        elements = None
-
-        if file_path.lower().endswith('.pdf'):
-            elements = partition_pdf(
-                filename=file_path,
-                strategy=strategy,
-                infer_table_structure=True,
-                model_name=model_name
-            )
-        elif file_path.lower().endswith(('.jpg', '.jpeg', '.png')):
-            elements = partition_image(
-                filename=file_path,
-                strategy=strategy,
-                infer_table_structure=True,
-                model_name=model_name
-            )
-
-        return elements
-
-    def load_text_data(self, elements, file_path, extract_tables):
-        elements_to_json(elements, filename=file_path)
-        text_file = self.process_json_file(file_path, extract_tables)
-
-        with open(text_file, 'r') as file:
-            content = file.read()
-
-        return content
-
-    def process_json_file(self, input_data, extract_tables):
-        # Read the JSON file
-        with open(input_data, 'r') as file:
-            data = json.load(file)
-
-        # Iterate over the JSON data and extract required table elements
-        extracted_elements = []
-        for entry in data:
-            if entry["type"] == "Table":
-                extracted_elements.append(entry["metadata"]["text_as_html"])
-            elif entry["type"] == "Title" and extract_tables is False:
-                extracted_elements.append(entry["text"])
-            elif entry["type"] == "NarrativeText" and extract_tables is False:
-                extracted_elements.append(entry["text"])
-            elif entry["type"] == "UncategorizedText" and extract_tables is False:
-                extracted_elements.append(entry["text"])
-
-        # Write the extracted elements to the output file
-        new_extension = 'txt'  # You can change this to any extension you want
-        new_file_path = self.change_file_extension(input_data, new_extension)
-        with open(new_file_path, 'w') as output_file:
-            for element in extracted_elements:
-                output_file.write(element + "\n\n")  # Adding two newlines for separation
-
-        return new_file_path
-
-    def change_file_extension(self, file_path, new_extension):
-        # Check if the new extension starts with a dot and add one if not
-        if not new_extension.startswith('.'):
-            new_extension = '.' + new_extension
-
-        # Split the file path into two parts: the base (everything before the last dot) and the extension
-        # If there's no dot in the filename, it'll just return the original filename without an extension
-        base = file_path.rsplit('.', 1)[0]
-
-        # Concatenate the base with the new extension
-        new_file_path = base + new_extension
-
-        return new_file_path
 
     # Function to safely evaluate type strings
     def safe_eval_type(self, type_str, context):
